move old stuff into old/ folder (move components as well). Extract icons
