+++ src/components/ProjectFilter.svelte
@@ -1,7 +1,10 @@
 <script>
-  export let viewMode = 'list'; // 'list' or 'grid'
+  import MenuIcon from "../icons/MenuIcon.svelte";
+  import GridIcon from "../icons/GridIcon.svelte";
+
+  export let viewMode = "list"; // 'list' or 'grid'
   export let onViewChange = () => {};
-  
+
   function setViewMode(mode) {
     viewMode = mode;
     onViewChange(mode);
@@ -12,31 +15,22 @@
   <div class="filter-content">
     <h2 class="projects-title">Projects</h2>
     <div class="view-toggle">
-    <button 
-      class="toggle-btn"
-      class:active={viewMode === 'list'}
-      on:click={() => setViewMode('list')}
-      aria-label="List view"
-    >
-      <svg width="16" height="16" viewBox="0 0 16 16" fill="currentColor">
-        <rect x="1" y="3" width="14" height="1.5"/>
-        <rect x="1" y="7" width="14" height="1.5"/>
-        <rect x="1" y="11" width="14" height="1.5"/>
-      </svg>
-    </button>
-    <button 
-      class="toggle-btn"
-      class:active={viewMode === 'grid'}
-      on:click={() => setViewMode('grid')}
-      aria-label="Grid view"
-    >
-      <svg width="16" height="16" viewBox="0 0 16 16" fill="currentColor">
-        <rect x="1" y="1" width="6" height="6" rx="1"/>
-        <rect x="9" y="1" width="6" height="6" rx="1"/>
-        <rect x="1" y="9" width="6" height="6" rx="1"/>
-        <rect x="9" y="9" width="6" height="6" rx="1"/>
-      </svg>
-    </button>
+      <button
+        class="toggle-btn"
+        class:active={viewMode === "list"}
+        on:click={() => setViewMode("list")}
+        aria-label="List view"
+      >
+        <MenuIcon />
+      </button>
+      <button
+        class="toggle-btn"
+        class:active={viewMode === "grid"}
+        on:click={() => setViewMode("grid")}
+        aria-label="Grid view"
+      >
+        <GridIcon />
+      </button>
     </div>
   </div>
 </div>
@@ -95,9 +89,5 @@
   .toggle-btn.active {
     background: var(--hover-bar-active);
     color: white;
-  }
-
-  .toggle-btn i {
-    font-size: 14px;
   }
 </style>
