+++ src/icons/GridIcon.svelte
@@ -1,0 +1,6 @@
+<svg width="16" height="16" viewBox="0 0 16 16" fill="currentColor">
+  <rect x="1" y="1" width="6" height="6" rx="1" />
+  <rect x="9" y="1" width="6" height="6" rx="1" />
+  <rect x="1" y="9" width="6" height="6" rx="1" />
+  <rect x="9" y="9" width="6" height="6" rx="1" />
+</svg>
