+++ src/icons/MenuIcon.svelte
@@ -1,0 +1,5 @@
+<svg width="16" height="16" viewBox="0 0 16 16" fill="currentColor">
+  <rect x="1" y="3" width="14" height="1.5" />
+  <rect x="1" y="7" width="14" height="1.5" />
+  <rect x="1" y="11" width="14" height="1.5" />
+</svg>
