+++ src/old/components/HalfMdHalfImageCell.svelte
