+++ src/old/projects/AIGenerationProject.svelte
