+++ src/old/projects/SpeechToLLMProject.svelte
